Review fixes 2nd round
