+++ jdbc/src/main/java/net/snowflake/client/internal/core/arrow/cursor/CursorState.java
@@ -5,10 +5,15 @@ public final class CursorState {
   private int currentRow = -1;
   private boolean afterLast = false;
 
+  private int currentRowInBatch = -1;
+  private int currentBatchRowCount = 0;
+
   public void reset() {
     wasNull = false;
     currentRow = -1;
     afterLast = false;
+    currentRowInBatch = -1;
+    currentBatchRowCount = 0;
   }
 
   public boolean wasNull() {
@@ -33,5 +38,30 @@ public final class CursorState {
 
   public void setAfterLast() {
     this.afterLast = true;
+  }
+
+  public int getCurrentRowInBatch() {
+    return currentRowInBatch;
+  }
+
+  void incrementRowInBatch() {
+    currentRowInBatch++;
+  }
+
+  int getCurrentBatchRowCount() {
+    return currentBatchRowCount;
+  }
+
+  boolean hasNextRowInBatch() {
+    return currentRowInBatch + 1 < currentBatchRowCount;
+  }
+
+  void startNewBatch(int rowCount) {
+    this.currentBatchRowCount = rowCount;
+    this.currentRowInBatch = 0;
+  }
+
+  boolean hasLoadedBatch() {
+    return currentRowInBatch >= 0;
   }
 }
