+++ jdbc/src/test/java/net/snowflake/client/internal/core/arrow/cursor/ArrowBatchManagerTest.java
@@ -15,21 +15,20 @@ public class ArrowBatchManagerTest {
             new int[] {}, new int[] {10, 20}, new int[] {}, new int[] {30})) {
       ArrowResources resources = resourcesHolder.getResources();
       CursorState cursor = new CursorState();
-      BatchState batch = new BatchState();
       SchemaState schema = new SchemaState(resources.getActiveRoot());
-      ArrowBatchManager manager = new ArrowBatchManager(cursor, batch, resources, schema);
+      ArrowBatchManager manager = new ArrowBatchManager(cursor, resources, schema);
 
       assertTrue(manager.fetchNextRow());
-      assertEquals(0, batch.getCurrentRowInBatch());
-      assertEquals(2, batch.getCurrentBatchRowCount());
+      assertEquals(0, cursor.getCurrentRowInBatch());
+      assertEquals(2, cursor.getCurrentBatchRowCount());
       assertFalse(cursor.isAfterLast());
 
       assertTrue(manager.fetchNextRow());
-      assertEquals(1, batch.getCurrentRowInBatch());
+      assertEquals(1, cursor.getCurrentRowInBatch());
 
       assertTrue(manager.fetchNextRow());
-      assertEquals(0, batch.getCurrentRowInBatch());
-      assertEquals(1, batch.getCurrentBatchRowCount());
+      assertEquals(0, cursor.getCurrentRowInBatch());
+      assertEquals(1, cursor.getCurrentBatchRowCount());
 
       assertFalse(manager.fetchNextRow());
       assertTrue(cursor.isAfterLast());
